moved model serialisation code into separate header
